main idea/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
openpyxl==3.1.2
google-generativeai==0.3.1
requests==2.31.0
python-multipart==0.0.6
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.25.2
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
